requirements.txt
google-generativeai

pypdf
python-dotenv
